Create sp_daily_build_prep_non_pro_tables
Program to create the daily build version of the non pro tables.  This does all of our validation and cleanup on those tables so that the stats and play info to follow is working off of the cleaned up tables instead of having to clean them up as they go.  Takes out wonky player ids, fixes incorrect player ids, etc.